fix tests and move to integration
