+++ packages/medusa/package.json
@@ -40,7 +40,8 @@
     "watch": "tsc --build --watch",
     "build": "rimraf dist && tsc --build",
     "serve": "node dist/app.js",
-    "test": "jest --silent=false --bail --maxWorkers=50% --forceExit"
+    "test": "jest --silent=false --bail --maxWorkers=50% --forceExit -- src/commands/plugin/db/__tests__/plugin-generate.spec.ts",
+    "test:integration": "jest --forceExit"
   },
   "devDependencies": {
     "@medusajs/framework": "^2.2.0",
