+++ packages/medusa/src/commands/plugin/db/integration-tests/__fixtures__/plugins-1-no-default/src/modules/module-1/models/module-model-1.ts
