+++ packages/medusa/src/commands/plugin/db/integration-tests/__fixtures__/plugins-1/src/modules/module-1/index.ts
