+++ packages/medusa/src/commands/plugin/db/integration-tests/__tests__/plugin-generate.spec.ts
@@ -1,7 +1,7 @@
 import { logger } from "@medusajs/framework/logger"
 import { FileSystem } from "@medusajs/framework/utils"
 import { join } from "path"
-import main from "../generate"
+import main from "../../generate"
 
 jest.mock("@medusajs/framework/logger")
 
@@ -36,7 +36,6 @@ describe("plugin-generate", () => {
         directory: join(__dirname, "..", "__fixtures__", "plugins-1"),
       })
 
-      console.log(logger.info["mock"])
       expect(logger.info).toHaveBeenNthCalledWith(1, "Generating migrations...")
       expect(logger.info).toHaveBeenNthCalledWith(
         2,
@@ -61,7 +60,6 @@ describe("plugin-generate", () => {
         directory: join(__dirname, "..", "__fixtures__", "plugins-1"),
       })
 
-      console.log(logger.info["mock"])
       expect(logger.info).toHaveBeenNthCalledWith(1, "Generating migrations...")
       expect(logger.info).toHaveBeenNthCalledWith(
         2,
